Added routines group of functions to modularize components folder.
