+++ Firmware_C3/src/components/DISPLAY/constants.h
@@ -1,7 +1,13 @@
+//* Paramethers for OLED display
+
 #define LINE0 0
 #define LINE1 16
 #define LINE2 32
 #define LINE3 48
+
+#define SCREEN_WIDTH 128
+#define SCREEN_HEIGHT 64
+#define SCREEN_ADDRESS 0x3C
 
 #define XCENTER 64
 #define YCENTER 32
