+++ Firmware_C3/src/components/DISPLAY/features/display.h
@@ -10,56 +10,6 @@
 #define _width 128
 #define _height 64
 
-void updateDisplay()
-{
-  int remainingTime = (TIME_TO_SLEEP - idleTimer) / SAMPLING_TIMER;
-  idleMinutes = remainingTime / 600;
-  idleSeconds = (remainingTime % 600) / 10;
-
-  display.clearDisplay();
-  display.setTextSize(1); // Set text size to 2 (you can adjust the size as needed)
-  display.setTextColor(SSD1306_WHITE);
-
-  display.setCursor(0, LINE0);
-  display.print("Temperature: ");
-  if (heaterTemperature < 100)
-    display.print(" ");
-  display.print(heaterTemperature);
-  display.print((char)247);
-  display.print("C");
-
-  display.setCursor(0, LINE1); // Adjust vertical position
-  display.print("Temp. Goal:  ");
-
-  if (tempGoal < 100)
-    display.print(" 0");
-  display.print(tempGoal);
-  display.print((char)247);
-  display.print("C");
-
-  display.setCursor(0, LINE2); // Adjust vertical position
-  display.print("PWM: ");
-  display.print(powerPercent);
-  display.print("%");
-  display.setCursor(92, LINE2);
-  display.print(battery);
-  display.print("V");
-  /*
-  display.setCursor(102, LINE2);
-  display.print(idleMinutes);
-  display.print(":");
-  if (idleSeconds < 10)
-    display.print("0");
-  display.print(idleSeconds);
-  */
-  // Display the IP address
-  display.setCursor(0, LINE3); // Adjust vertical position
-  display.print("IP: ");
-  display.print(WiFi.localIP());
-
-  display.display();
-}
-
 void drawXbm(const uint8_t *xbm_data, int16_t width, int16_t height, int16_t x, int16_t y)
 {
   display.drawXBitmap(x, y, xbm_data, width, height, SSD1306_WHITE);
