--- Firmware_C3/src/components/PUSHBUTTON/docs/docs.txt
@@ -1,4 +1,0 @@
-
-    ---> Datasheet:
-    ---> Módulo:
-    ---> Tutorial:
