--- Firmware_C3/src/components/PUSHBUTTON/setup.h
@@ -1,8 +1,0 @@
-#if !defined(SETUP_PUSHBUTTON)
-#define SETUP_PUSHBUTTON
-
-#include "shared/dependencies.h"
-// ADICIONAR AQUI OS IMPORTS DOS ARQUIVOS DAS FEATURES.
-#include "components/PUSHBUTTON/features/clicky.h"
-
-#endif // PUSHBUTTON
