--- Firmware_C3/src/components/THERMISTOR/setup.h
@@ -1,7 +1,0 @@
-#if !defined(SETUP_THERMISTOR)
-#define SETUP_THERMISTOR
-
-#include "shared/dependencies.h"
-#include "components/THERMISTOR/features/read_temp_thermistor.h"
-
-#endif // SETUP_THERMISTOR
