+++ Firmware_C3/src/main.cpp
@@ -1,10 +1,6 @@
 //* ---------------------- IMPORTS ----------------------
 //* Shared:
 #include "shared/dependencies.h"
-//* Utilities:
-//* Components:
-#include "components/PUSHBUTTON/setup.h"
-#include "components/DISPLAY/setup.h"
 
 void setup()
 {
@@ -21,37 +17,20 @@ void loop()
 {
   ArduinoOTA.handle();
 
-  if (digitalRead(buttonPin))
-  {
-    int shutdownCounter=0;
-    while (digitalRead(buttonPin))
-    {
-      sampleRandomLED();
-      shutdownCounter++;
-      if (shutdownCounter>100)
-      {
-        digitalWrite(latchPin, LOW);
-        setLED(0,0,0);
-        esp_deep_sleep_start();
-      }
-      delay(25);
-    }
-    idleTimer = 0;
-  }
+  buttonPress();
 
   if ((logTimer > SAMPLES_TO_SEC)) // logs variables every 1s if awake
   {
     logTimer = 0;
-    thermistor = calculate_resistance();
-    heaterTemperature = steinhart(thermistor);
     battery = analogReadMilliVolts(1) * 8.2 / 4095;
-    //Serial.print(".");
     updateDisplay();
     sampleRandomLED();
   }
 
   if ((millis() - globalTimer) > SAMPLING_TIMER)
   {
+    thermistor = calculate_resistance();
+    heaterTemperature = steinhart(thermistor);
     globalTimer = millis();
     idleTimer++;
     logTimer++;
@@ -60,8 +39,8 @@ void loop()
 
   if (idleTimer > (240 * SAMPLES_TO_SEC))
   {
-    setLED(0, 0, 0);
-    //digitalWrite(latchPin, LOW);
-    //esp_deep_sleep_start();
+    // setLED(0, 0, 0);
+    // digitalWrite(latchPin, LOW);
+    // esp_deep_sleep_start();
   }
 }
